+++ example/android/app/src/main/kotlin/com/rocksti/get_thumbnail_video_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.rocksti.get_thumbnail_video_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()